temporarily disabled auto-katakana conversion
The code to take romaji words using only a capital letter at the
beginning of the word and not all capital letters and output to
all katakana syllables was too inefficient and collision-prone.

Maybe in the future I will find time and reason to re-add it as
a more elegant implementation, but in the meantime, there's no
real problem with just writing in all caps for katakana.